move supporting pages to the component
