+++ pages/community.js
@@ -6,8 +6,8 @@ import teamcss from '../styles/teams.module.scss';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import teams from '../static/data/team_members.json';
-import Member from '../pages/member';
-import Team from '../pages/teamcard';
+import Member from '../components/member';
+import Team from '../components/teamcard';
 import Link from 'next/link';
 import Image from 'next/image';
 import FacebookIcon from '@material-ui/icons/Facebook';
